+++ pyccel/commands/pyccel_test.py
@@ -9,10 +9,9 @@ import json
 import os
 import pathlib
 import sys
-from argparse import ArgumentParser
 from importlib.metadata import Distribution
 
-from .argparse_helpers import add_help_flag, add_version_flag, deprecation_warning
+from .argparse_helpers import add_help_flag
 from .pyccel_clean import pyccel_clean
 
 __all__ = (
@@ -275,7 +274,7 @@ def pyccel_test(*, folder, dry_run, verbose, language, run_mpi):
     sys.exit(final_retcode)
 
 
-def setup_pyccel_test_parser(parser, add_version=False):
+def setup_pyccel_test_parser(parser):
     """
     Add the `pyccel test` arguments to the parser.
 
@@ -285,16 +284,10 @@ def setup_pyccel_test_parser(parser, add_version=False):
     ----------
     parser : argparse.ArgumentParser
         The parser to be modified.
-    add_version : bool, default=False
-        Indicates whether a --version flag should be added to the command.
-        This option will be removed in v2.3.
     """
     group = parser.add_argument_group("Options")
 
     add_help_flag(group)
-
-    if add_version:
-        add_version_flag(group)
 
     group.add_argument(
         "--dry-run",
@@ -331,32 +324,3 @@ def setup_pyccel_test_parser(parser, add_version=False):
         dest="run_mpi",
         help="Do not run the parallel tests.",
     )
-
-
-def pyccel_test_command():
-    """
-    Command line wrapper for the deprecated `pyccel-test` command line tool.
-
-    Command line wrapper for the deprecated `pyccel-test` command line tool.
-
-    Returns
-    -------
-    pytest.ExitCode
-        The pytest return code.
-    """
-    parser = ArgumentParser(
-        description="Tool for running the test suite of Pyccel", add_help=False
-    )
-
-    setup_pyccel_test_parser(parser, add_version=True)
-
-    print(deprecation_warning("test"), file=sys.stderr)
-
-    # Parse the command line arguments
-    args = parser.parse_args()
-
-    print()
-    retcode = pyccel_test(**vars(args))
-    print()
-
-    return retcode